split up fetching & data manipulation, refactor
